Update positions of switcher in both network and form editors.
