+++ src/styles/editor-type-switcher.css
@@ -1,0 +1,11 @@
+.editor-type-switcher {
+  font-size: 2.25em;
+  margin-right: 0.125em;
+  transition-property: filter;
+  transition-timing-function: ease-out;
+  transition-duration: 250ms;
+
+  &:active {
+    filter: brightness(50%);
+  }
+}
