+++ src/styles/editor/editor.css
@@ -73,7 +73,7 @@
   font-size: 1.25em;
 }
 
-.editor-title {
+.editor-header {
   z-index: 1;
   position: absolute;
   top: 0;
@@ -82,7 +82,7 @@
   pointer-events: none;
 }
 
-.editor-title-content {
+.editor-header-content {
   display: inline-block;
   background: rgba(255, 255, 255, 0.5);
   padding: 0.375em 0.5em;
