+++ src/styles/editor/undo-remove.css
@@ -1,6 +1,6 @@
 .editor-undo-rm {
   position: absolute;
-  top: 2.5em;
+  top: 3.7em;
   left: 50%;
   transform: translate(-50%, 0);
   text-align: center;
