+++ src/styles/index.css
@@ -30,3 +30,4 @@
 @import "./main-menu.css";
 @import "./reach-interval-highlighter.css";
 @import "./copy-field.css";
+@import "./editor-type-switcher.css";
